+++ keep_alive.py
@@ -3,7 +3,7 @@ from threading import Thread
 app = Flask('')
 @app.route('/')
 def main():
-  return "Your bot is alive!"
+  return "Hello! Merlin-py is here!"
 def run():
   app.run(host="0.0.0.0", port=8080)
 def keep_alive():